requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
gcloud==0.18.3
google-api-python-client==1.7.4
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauth2client==4.1.2
six==1.11.0
Werkzeug==0.12.2
